Moved a GameManager, main, timer and physics into their own folder. Fixed more "ECS" stuff. Added a player class, check the player class for sort of a template on how to create gameobjects with components. Physics collisions and rigidbodies haven't been tested yet.
