+++ src/components/Components.h
@@ -1,3 +1,4 @@
-#include "ECS.h"
+//This file should have includes to separate components
+//If a component includes this file we'll run into circular dependency problems
 #include "SpriteComponent.h"
 #include "RigidBodyComponent.h"
