+++ src/graphics/ShaderVert.glsl
@@ -6,9 +6,13 @@ layout (location = 2) in vec2 aTexCoord;
 
 out vec3 ourColor;
 out vec2 TexCoord;
+
+uniform mat4 modelMatrix;
+
 void main()
 {
-    gl_Position = vec4(aPos, 1.0);
     ourColor = aColor;
     TexCoord = aTexCoord;
+
+    gl_Position =  vec4(aPos, 1.0);
 }
