+++ src/scenes/Scene1.h
@@ -1,12 +1,13 @@
 #ifndef SCENE1_H
 #define SCENE1_H
 #include "Scene.h"
-#include "graphics/shader.h"
-#include "graphics/TextureManager.h"
+
+class Player;
+
 class Scene1 : public Scene
 {
 private:
-	TextureManager* texture;
+	Player* player;
 
 public:
 	Scene1();
